refactor: BookApi.java, PersonalBookRecord 분리
