+++ src/main/java/com/dokdok/book/controller/BookController.java
@@ -63,37 +63,5 @@ public class BookController implements BookApi {
         return ApiResponse.deleted("책 삭제 성공");
     }
 
-    @Override
-    @PostMapping("/{bookId}")
-    public ResponseEntity<ApiResponse<PersonalReadingRecordCreateResponse>> createMyReadingRecord(@PathVariable Long bookId, @RequestBody PersonalReadingRecordCreateRequest request) {
-        PersonalReadingRecordCreateResponse response = personalReadingRecordService.create(bookId, request);
-        return ApiResponse.created(response, "기록 등록 성공");
-    }
 
-    @Override
-    @PatchMapping("/{bookId}/records/{recordId}")
-    public ResponseEntity<ApiResponse<PersonalReadingRecordCreateResponse>> updateMyReadingRecord(@PathVariable Long bookId, @PathVariable Long recordId, @RequestBody PersonalReadingRecordUpdateRequest request) {
-        PersonalReadingRecordCreateResponse response = personalReadingRecordService.update(bookId, recordId, request);
-        return ApiResponse.success(response, "기록 수정 성공");
-    }
-
-    @Override
-    @DeleteMapping("/{bookId}/records/{recordId}")
-    public ResponseEntity<ApiResponse<Void>> deleteMyReadingRecord(@PathVariable Long bookId, @PathVariable Long recordId) {
-        personalReadingRecordService.delete(bookId, recordId);
-        return ApiResponse.deleted("기록 삭제 성공");
-    }
-
-    @Override
-    @GetMapping("/{bookId}/records")
-    public ResponseEntity<ApiResponse<PageResponse<PersonalReadingRecordListResponse>>> getMyReadingRecords(
-            @PathVariable Long bookId,
-            @PageableDefault(size = 10, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable
-    ) {
-
-        Page<PersonalReadingRecordListResponse> records = personalReadingRecordService.getRecords(bookId, pageable);
-        PageResponse<PersonalReadingRecordListResponse> response = PageResponse.from(records);
-        return ApiResponse.success(response, "기록 조회 성공");
-
-    }
 }
